add product function

+++ src/main/java/com/example/FarmMarket/repository/FarmMarketRepository.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
@@ -24,7 +25,7 @@ public class FarmMarketRepository {
     }
 
     public void newProduct(String category, String product_name, String product_description,
-                           BigInteger price, BigInteger amount) {
+                           BigDecimal price, BigDecimal amount) {
         String sql = "insert into product (category, product_name, product_description, price, amount) values " +
                 "(:category, :product_name, :product_description, :price, :amount)";
         Map<String, Object> paramMap = new HashMap<>();
